daemon/requirements.txt
# Faramesh Guard Daemon
# Python dependencies

# Web framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Async HTTP
aiohttp>=3.9.0
httpx>=0.26.0

# Database
aiosqlite>=0.19.0

# Security
cryptography>=41.0.0
PyJWT>=2.8.0

# Policy engine (optional - can use built-in rules)
# opa-python>=1.0.0  # Uncomment for OPA/Rego support

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Monitoring
psutil>=5.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
